start updating vgamepad again
