+++ game_tools/src/game_core.py
@@ -79,16 +79,6 @@ _key_up_pending_jobs = {}
 _camera_speed = None
 _camera_snap_angle = None
 _game_use_awsd_for_arrows = False
-xbox_power = {
-    "left_stick": 1,
-    "right_stick": 1,
-    "left_trigger": 1,
-    "right_trigger": 1,
-}
-# _xbox_left_analog_power = 1
-# _xbox_right_analog_power = 1
-# _xbox_left_trigger_power = 1
-# _xbox_right_trigger_power = 1
 
 DIR_MODE_CAM_CONTINUOUS = "continuous"
 DIR_MODE_CAM_SNAP = "snap"
@@ -110,12 +100,6 @@ EVENT_ON_MOUSE = "on_mouse"
 EVENT_MOUSE_CLICK = "click"
 EVENT_MOUSE_HOLD = "hold"
 EVENT_MOUSE_RELEASE = "release"
-
-@dataclass
-class GameXboxEvent:
-    subject: str
-    type: str
-    value: Any
 
 def no_op():
     pass
@@ -581,75 +565,6 @@ def game_gear_set(gear_num: int):
         angle = settings.get("user.game_camera_snap_gear_angles").split(" ").get(gear_num)
         camera_snap_dynamic_set_angle(angle)
 
-xbox_dir_hold_left_analog_dir_map = {
-    "up": lambda power: actions.user.vgamepad_left_joystick(0, power),
-    "down": lambda power: actions.user.vgamepad_left_joystick(0, -power),
-    "left": lambda power: actions.user.vgamepad_left_joystick(-power, 0),
-    "right": lambda power: actions.user.vgamepad_left_joystick(power, 0),
-}
-
-xbox_dir_hold_right_analog_dir_map = {
-    "up": lambda power: actions.user.vgamepad_right_joystick(0, power),
-    "down": lambda power: actions.user.vgamepad_right_joystick(0, -power),
-    "left": lambda power: actions.user.vgamepad_right_joystick(-power, 0),
-    "right": lambda power: actions.user.vgamepad_right_joystick(power, 0),
-}
-
-xbox_button_map = {
-    "a": "a",
-    "b": "b",
-    "x": "x",
-    "y": "y",
-    "dpad_up": "dpad_up",
-    "dpad_down": "dpad_down",
-    "dpad_left": "dpad_left",
-    "dpad_right": "dpad_right",
-    "lb": "left_shoulder",
-    "rb": "right_shoulder",
-    "lt": "left_trigger",
-    "rt": "right_trigger",
-    "l1": "left_shoulder",
-    "r1": "right_shoulder",
-    "l2": "left_trigger",
-    "r2": "right_trigger",
-    "l3": "left_thumb",
-    "r3": "right_thumb",
-    "left_shoulder": "left_shoulder",
-    "right_shoulder": "right_shoulder",
-    "left_thumb": "left_thumb",
-    "right_thumb": "right_thumb",
-    "left_trigger": "left_trigger",
-    "right_trigger": "right_trigger",
-    "start": "start",
-    "back": "back",
-    "guide": "guide",
-}
-
-def xbox_dir_hold_left_analog(dir: str, power: float = None):
-    """Hold a left analog direction"""
-    xbox_dir_hold_left_analog_dir_map[dir](power or xbox_power["left_stick"])
-
-def xbox_dir_hold_right_analog(dir: str, power: float = None):
-    """Hold a right analog direction"""
-    xbox_dir_hold_right_analog_dir_map[dir](power or xbox_power["right_stick"])
-
-def xbox_dir_hold_dpad(dir: str):
-    """Hold a dpad direction"""
-    actions.user.vgamepad_button("dpad_up", up=True)
-    actions.user.vgamepad_button("dpad_down", up=True)
-    actions.user.vgamepad_button("dpad_left", up=True)
-    actions.user.vgamepad_button("dpad_right", up=True)
-    actions.user.vgamepad_button(f"dpad_{dir}", down=True)
-
-def xbox_set_power(subject: str, power: Union[str, int, float]):
-    power = float(power)
-    if subject not in xbox_power:
-        raise ValueError(f"xbox_set_power subject must be one of {xbox_power.keys()}. Got {subject}")
-    if power < 0 or power > 1:
-        raise ValueError(f"xbox_set_power power must be between 0 and 1.0. Got {power}")
-    xbox_power[subject] = power
-    actions.user.game_event_trigger_on_xbox_gamepad_event(subject, "power", power)
-
 @mod.action_class
 class Actions:
     def game_calibrate_x_360_add(num: int):
@@ -767,39 +682,6 @@ class Actions:
         if EVENT_ON_MOUSE in event_subscribers:
             for callback in event_subscribers[EVENT_ON_MOUSE]:
                 callback(button, state)
-
-    def game_event_register_on_xbox_gamepad_event(callback: callable):
-        """
-        ```
-        def on_gamepad_event(event: Any):
-            print(event.subject, event.event_type, event.value)
-
-        actions.user.game_event_register_on_xbox_gamepad_event(on_gamepad_event)
-        ```
-        """
-        global event_subscribers
-        if "on_xbox" not in event_subscribers:
-            event_subscribers["on_xbox"] = []
-        event_subscribers["on_xbox"].append(callback)
-
-    def game_event_unregister_on_xbox_gamepad_event(callback: callable):
-        """
-        Unregister a callback for a specific game event.
-        """
-        global event_subscribers
-        if "on_xbox" in event_subscribers:
-            event_subscribers["on_xbox"].remove(callback)
-            if not event_subscribers["on_xbox"]:
-                del event_subscribers["on_xbox"]
-
-    def game_event_trigger_on_xbox_gamepad_event(subject: str, type: str, value: Any):
-        """
-        Trigger an event and call all registered callbacks.
-        """
-        global event_subscribers
-        if "on_xbox" in event_subscribers:
-            for callback in event_subscribers["on_xbox"]:
-                callback(GameXboxEvent(subject, type, value))
 
     def game_event_unregister_all():
         """Unregister all game events"""
